nodeTransformController 20%

+++ ToEnd/CGHGraphicResource.cpp
@@ -1,17 +1,17 @@
 #include "CGHGraphicResource.h"
 
 std::vector<std::string> TextureInfo::s_textureNames;
-std::unordered_map<std::string, size_t> TextureInfo::s_textureIndex;
+std::unordered_map<std::string, unsigned int> TextureInfo::s_textureIndex;
 
-size_t TextureInfo::GetTextureFilePathID(const char* path)
+unsigned int TextureInfo::GetTextureFilePathID(const char* path)
 {
-	size_t result = 0;
+	unsigned int result = 0;
 
 	auto iter = s_textureIndex.find(path);
 
 	if (iter == s_textureIndex.end())
 	{
-		result = s_textureIndex.size();
+		result = static_cast<unsigned int>(s_textureIndex.size());
 		s_textureNames.push_back(path);
 		s_textureIndex[path] = result;
 	}
